# apps/api/storage/framework/testing-commerce-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Watch",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "security": [
        {
            "sessionCookie": []
        }
    ],
    "paths": {
...
        "/api/v1/admin/orders/{id}/status": {
            "patch": {
                "operationId": "order.updateStatus",
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated order.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AdminOrderResource"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "csrfHeader": {
                "type": "apiKey",
                "description": "CSRF token from the XSRF-TOKEN cookie (Sanctum SPA).",
                "in": "header",
                "name": "X-XSRF-TOKEN"
            },
            "sessionCookie": {
                "type": "apiKey",
                "description": "Sanctum SPA session cookie.",
                "in": "cookie",
                "name": "laravel_session"
            }
        },
        "schemas": {
            "AdminOrderResource": {
                "type": "object",
                "properties": {
                    "customer_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "id": {
                        "type": "integer"
                    },
                    "number": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "subtotal": {
                        "type": "string"
                    },
                    "discount_total": {
                        "type": "string"
                    },
                    "shipping_total": {
                        "type": "string"
                    },
                    "tax_total": {
                        "type": "string"
                    },
                    "grand_total": {
                        "type": "string"
                    },
                    "shipping_address": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "properties": {
                            "recipient_name": {
                                "type": "string"
                            },
                            "phone": {
                                "type": "string"
                            },
                            "province_code": {
                                "type": "string"
                            },
                            "district_code": {
                                "type": "string"
                            },
                            "ward_code": {
                                "type": "string"
                            },
                            "address_line": {
                                "type": "string"
                            },
                            "postal_code": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "recipient_name",
                            "phone",
                            "province_code",
                            "district_code",
                            "ward_code",
                            "address_line",
                            "postal_code"
                        ]
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "status_history": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderStatusHistory"
                        }
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "customer_id",
                    "id",
                    "number",
                    "status",
                    "currency",
                    "subtotal",
                    "discount_total",
                    "shipping_total",
                    "tax_total",
                    "grand_total",
                    "shipping_address",
                    "created_at"
                ],
                "title": "AdminOrderResource"
            },
...
            "OrderItem": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "order_id": {
                        "type": "integer"
                    },
                    "product_variant_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "sku": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "qty": {
                        "type": "integer"
                    },
                    "unit_price": {
                        "type": "string"
                    },
                    "line_total": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "order_id",
                    "product_variant_id",
                    "sku",
                    "name",
                    "qty",
                    "unit_price",
                    "line_total",
                    "created_at",
                    "updated_at"
                ],
                "title": "OrderItem"
            },
...
            "OrderStatusHistory": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "order_id": {
                        "type": "integer"
                    },
                    "from_status": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "to_status": {
                        "type": "string"
                    },
                    "changed_by_admin_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "changed_by_customer_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "note": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "order_id",
                    "from_status",
                    "to_status",
                    "changed_by_admin_id",
                    "changed_by_customer_id",
                    "note",
                    "created_at",
                    "updated_at"
                ],
                "title": "OrderStatusHistory"
            },
...
            "UpdateOrderStatusRequest": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "paid",
                            "fulfilling",
                            "shipped",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "note": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 1000
                    }
                },
                "required": [
                    "status"
                ],
                "title": "UpdateOrderStatusRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}